// packages/amazon-connect-contact-flow-paths-finder/src/contact-flow/enums/action-blocks/contact-flowaction-block-types.ts
export enum ContactFlowActionBlockTypes {
  PLAY_PROMPT = "MessageParticipant",
  AMAZON_Q_IN_CONNECT = "CreateWisdomSession",
  AUTHENTICATE_CUSTOMER = "AuthenticateParticipant",
  CHECK_CONTACT_ATTRIBUTES = "Compare",
  CHECK_HOURS_OF_OPERATION = "CheckHoursOfOperation",
  CHECK_QUEUE_STATUS = "CheckMetricData",
  CHECK_STAFFING = "CheckMetricData",
  CONTACT_TAGS = "TagContact",
  UNTAG_CONTACT = "UnTagContact",
  GET_PROFILE = "GetCustomerProfile",
  CREATE_PROFILE = "CreateCustomerProfile",
  UPDATE_PROFILE = "UpdateCustomerProfile",
  GET_PROFILE_OBJECT = "GetCustomerProfileObject",
  GET_CALCULATED_ATTRIBUTES = "GetCalculatedAttributesForCustomerProfile",
  ASSOCIATE_CONTACT_TO_PROFILE = "AssociateContactToCustomerProfile",
  DISCONNECT_PARTICIPANT = "DisconnectParticipant",
  DISTRIBUTE_BY_PERCENTAGE = "DistributeByPercentage",
  END_FLOW_EXECUTION = "EndFlowExecution",
  GET_CUSTOMER_INPUT = "GetParticipantInput",
  GET_QUEUE_METRICS = "GetMetricData",
  HOLD_CUSTOMER_OR_AGENT = "UpdatePreviousContactParticipantState",
  INVOKE_LAMBDA_FUNCTION = "InvokeLambdaFunction",
  INVOKE_FLOW_MODULE = "InvokeFlowModule",
  LOOP = "Loop",
  SET_CONTACT_ATTRIBUTES = "UpdateContactAttributes",
  SET_FLOW_ATTRIBUTES = "UpdateFlowAttributes",
  SET_CUSTOMER_QUEUE_FLOW = "UpdateContactEventHooks",
  SET_DISCONNECT_FLOW = "UpdateContactEventHooks",
  SET_LOGGING_BEHAVIOR = "UpdateFlowLoggingBehavior",
  SET_RECORDING_AND_ANALYTICS_BEHAVIOR = "UpdateContactRecordingBehavior",
  SET_ROUTING_CRITERIA = "UpdateRoutingCriteria",
  SET_VOICE = "UpdateContactTextToSpeechVoice",
  SET_WHISPER_FLOW = "UpdateContactEventHooks",
  SET_WORKING_QUEUE = "UpdateContactTargetQueue",
  SHOW_VIEW = "ShowView",
  STORE_CUSTOMER_INPUT = "GetParticipantInput",
  TRANSFER_TO_AGENT = "TransferContactToAgent",
  TRANSFER_TO_FLOW = "TransferToFlow",
  TRANSFER_TO_PHONE_NUMBER = "TransferParticipantToThirdParty",
  TRANSFER_TO_QUEUE = "TransferContactToQueue",
  WAIT = "Wait",
}